chore(effect): runs `pnpm workspace:rm`
