--- packages/schema-effect-adapter/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@traversable/schema-effect-adapter",
-  "type": "module",
-  "version": "0.0.0",
-  "private": false,
-  "description": "",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/traversable/schema.git",
-    "directory": "packages/schema-effect-adapter"
-  },
-  "bugs": {
-    "url": "https://github.com/traversable/schema/issues",
-    "email": "ahrjarrett@gmail.com"
-  },
-  "@traversable": {
-    "generateExports": { "include": ["**/*.ts"] },
-    "generateIndex": { "include": ["**/*.ts"] }
-  },
-  "publishConfig": { "access": "public" },
-  "scripts": {
-    "bench": "echo NOTHING TO BENCH",
-    "build": "pnpm build:esm && pnpm build:cjs && pnpm build:annotate",
-    "build:annotate": "babel build --plugins annotate-pure-calls --out-dir build --source-maps",
-    "build:esm": "tsc -b tsconfig.build.json",
-    "build:cjs": "babel build/esm --plugins @babel/transform-export-namespace-from --plugins @babel/transform-modules-commonjs --out-dir build/cjs --source-maps",
-    "check": "tsc -b tsconfig.json",
-    "clean": "pnpm run \"/^clean:.*/\"",
-    "clean:build": "rm -rf .tsbuildinfo dist build",
-    "clean:deps": "rm -rf node_modules",
-    "test": "vitest"
-  },
-  "peerDependencies": {
-    "@traversable/json": "workspace:^",
-    "@traversable/registry": "workspace:^"
-  },
-  "devDependencies": {
-    "@traversable/json": "workspace:^",
-    "@traversable/registry": "workspace:^"
-  }
-}
