--- packages/schema-effect-adapter/src/exports.ts
@@ -1,1 +1,0 @@
-export * from './version.js'
