--- packages/schema-effect-adapter/src/index.ts
@@ -1,2 +1,0 @@
-export * from './exports.js'
-export * as schemaEffectAdapter from './exports.js'
