--- packages/schema-effect-adapter/src/version.ts
@@ -1,3 +1,0 @@
-import pkg from './__generated__/__manifest__.js'
-export const VERSION = `${pkg.name}@${pkg.version}` as const
-export type VERSION = typeof VERSION
