--- packages/schema-effect-adapter/tsconfig.build.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.src.json",
-  "compilerOptions": {
-    "tsBuildInfoFile": ".tsbuildinfo/build.tsbuildinfo",
-    "types": ["node"],
-    "declarationDir": "build/dts",
-    "outDir": "build/esm",
-    "stripInternal": true
-  },
-  "references": [{ "path": "../json" }, { "path": "../registry" }]
-}
