--- packages/schema-effect-adapter/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "include": [],
-  "references": [
-    { "path": "tsconfig.src.json" },
-    { "path": "tsconfig.test.json" }
-  ]
-}
