--- packages/schema-effect-adapter/tsconfig.src.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "tsBuildInfoFile": ".tsbuildinfo/src.tsbuildinfo",
-    "rootDir": "src",
-    "types": ["node"],
-    "outDir": "build/src"
-  },
-  "references": [{ "path": "../json" }, { "path": "../registry" }],
-  "include": ["src"]
-}
