--- packages/schema-effect-adapter/tsconfig.test.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "tsBuildInfoFile": ".tsbuildinfo/test.tsbuildinfo",
-    "rootDir": "test",
-    "types": ["node"],
-    "noEmit": true
-  },
-  "references": [
-    { "path": "tsconfig.src.json" },
-    { "path": "../json" },
-    { "path": "../registry" }
-  ],
-  "include": ["test"]
-}
