--- packages/schema-effect-adapter/vite.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig, mergeConfig } from 'vitest/config'
-import sharedConfig from '../../vite.config.js'
-
-const localConfig = defineConfig({})
-
-export default mergeConfig(sharedConfig, localConfig)
