+++ tsconfig.json
@@ -6,7 +6,6 @@
     { "path": "packages/registry" },
     { "path": "packages/schema" },
     { "path": "packages/schema-core" },
-    { "path": "packages/schema-effect-adapter" },
     { "path": "packages/schema-seed" },
     { "path": "packages/schema-valibot-adapter" },
     { "path": "packages/schema-zod-adapter" }
